feat: :sparkles: make people able to be Stunned because of being attacked
